move object.is -> object-is
