+++ types/object-is/.npmignore
