+++ types/object-is/auto.d.ts
