+++ types/object-is/package.json
@@ -1,12 +1,12 @@
 {
     "private": true,
-    "name": "@types/object.is",
-    "version": "1.0.9999",
+    "name": "@types/object-is",
+    "version": "1.1.9999",
     "projects": [
-        "https://github.com/m59peacemaker/js-object.is#readme"
+        "https://github.com/es-shims/object-is#readme"
     ],
     "devDependencies": {
-        "@types/object.is": "workspace:."
+        "@types/object-is": "workspace:."
     },
     "owners": [
         {
